+++ lib/http/request.rb
@@ -178,29 +178,9 @@ module HTTP
     # @return [HTTP::Request]
     # @api public
     def redirect(uri, verb = @verb)
-      headers = self.headers.dup
-      headers.delete(Headers::HOST)
-
       redirect_uri = @uri.join(uri)
-
-      # Strip sensitive auth headers when redirecting to a different origin
-      # (scheme + host + port) to prevent credential leakage.
-      # See: https://github.com/httprb/http/issues/770
-      headers.delete(Headers::AUTHORIZATION) unless @uri.origin == redirect_uri.origin
-
-      new_body = body.source
-      if verb == :get
-        # request bodies should not always be resubmitted when following a redirect
-        # some servers will close the connection after receiving the request headers
-        # which may cause Errno::ECONNRESET: Connection reset by peer
-        # see https://github.com/httprb/http/issues/649
-        # new_body = Request::Body.new(nil)
-        new_body = nil
-        # the CONTENT_TYPE header causes problems if set on a get request w/ an empty body
-        # the server might assume that there should be content if it is set to multipart
-        # rack raises EmptyContentError if this happens
-        headers.delete(Headers::CONTENT_TYPE)
-      end
+      headers = redirect_headers(redirect_uri, verb)
+      new_body = verb == :get ? nil : body.source
 
       self.class.new(
         verb:           verb,
@@ -303,6 +283,31 @@ module HTTP
     end
 
     private
+
+    # Build headers for a redirect request
+    #
+    # Strips the Host header (it will be regenerated), sensitive credentials
+    # on cross-origin redirects, and Content-Type on GET verb changes.
+    #
+    # @param [HTTP::URI] redirect_uri the target URI
+    # @param [Symbol] verb the HTTP verb for the redirected request
+    # @return [HTTP::Headers] headers for the redirect
+    # @api private
+    def redirect_headers(redirect_uri, verb)
+      headers = self.headers.dup
+      headers.delete(Headers::HOST)
+
+      # Strip sensitive headers when redirecting to a different origin
+      # (scheme + host + port) to prevent credential leakage.
+      unless @uri.origin == redirect_uri.origin
+        headers.delete(Headers::AUTHORIZATION)
+        headers.delete(Headers::COOKIE)
+      end
+
+      headers.delete(Headers::CONTENT_TYPE) if verb == :get
+
+      headers
+    end
 
     # @!attribute [r] host
     #   Host from the URI